LayoutTests/imported/w3c/web-platform-tests/css/css-tables/visibility-collapse-row-004-expected.txt

Visibility collapse changes table height even when its height is fixed

a
b

FAIL collapsed row shrinks table height assert_equals: expected 100 but got 150
